+++ packages/jest-cli/src/init/questions.ts
@@ -25,6 +25,16 @@ const defaultQuestions: Array<PromptObject> = [
     type: 'confirm',
   },
   {
+    choices: [
+      {title: 'v8', value: 'v8'},
+      {title: 'babel', value: 'babel'},
+    ],
+    initial: 0,
+    message: 'Which provider should be used to instrument code for coverage?',
+    name: 'coverageProvider',
+    type: 'select',
+  },
+  {
     initial: false,
     message: 'Automatically clear mock calls and instances between every test?',
     name: 'clearMocks',